requirements.txt
langchain
langchain-anthropic
langchain-aws
langchain-community
chromadb
pypdf
python-dotenv
streamlit
faiss-cpu
tiktoken
